requirements.txt
streamlit==1.31.0
streamlit-lottie==0.0.5
streamlit-paste-button==0.1.2
openai==1.12.0
python-dotenv==1.0.1
pillow>=10.2.0
PyPDF2==3.0.1
requests==2.31.0
youtube-transcript-api==0.6.2
beautifulsoup4==4.12.3
python-docx==1.1.0
lxml==5.1.0
html5lib==1.1
gTTS==2.5.0
moviepy==1.0.3
reportlab==4.0.9
markdown==3.5.2
Pillow==10.2.0
